False commit to write about previous commit
Now, our code writes a file, in JSON format, with the data that we
would like to store.
to make the JSON stringily function work, the object passes to be
turned into JSON can’t have null fields in his structure. This is what
happens when you set properties, in an array object, not using the
array index to access it. What I did was to change the object creation
from “var obj = []“ to “var “obj = {}”. Now it generates pairs
“<propertyName> : propertyValue>” and this can be seen in the JSON file.