file where all data is store
this file is written in a single line, but we can make it look
prettier, with new lines and paragraph. It’s not necessary but if we
want to read the file with our own eyes, it helps us a lot when it
looks pretty. So i have set the code to write everything in a single
line, because I believe it writes faster and will make the read run
faster too.