code now is requesting and treating gzip responses
changed the request header to make a request that will return the
response in gzip compression. i also had to decompress this gzip’ed
response.
this makes our response be received around 2 times faster that it used
to be on flat json.
the response is a json file compressed in gzip