reorganized try catch blocks, saving last update
- i’ve changed whats inside the try catch block. it only try’s to parse
the json and if it fails. i set our json to null and added and if
statement check if json is not null.
- i am adding the last date, and time, when we get a good response from
dados abertos. I add this date to the data being passed to the parent
thread. just so we can send this information on our responses. this is
useful to inform users how old is the data they are seeing, as our
services keeps serving them even when dadosabertos server can’t serve
data.