watching log files instead of retrieving it again
instead of retrieving the end of each file to send the last lines on
response, now we keep track of the last lines on memory. we get the
last lines once, than we watch for changes on file and add new lines to
memory, removing old ones. keeping always the most recent ones.